+++ src/Sinch/Verification/SinchEvents/VerificationStartEventResponse.cs
@@ -1,13 +1,14 @@
 using System.Text.Json.Serialization;
 
-namespace Sinch.Verification.Hooks
+namespace Sinch.Verification.SinchEvents
 {
-    public class RequestEventResponseBase
+    public abstract class VerificationStartEventResponse
     {
         /// <summary>
         ///     Determines whether the verification can be executed.
         /// </summary>
         [JsonPropertyName("action")]
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         public Action? Action { get; set; }
     }
 }